+++ lib/src/framework/core/components/navigation/routing/hover_route.dart
@@ -2,5 +2,5 @@ import 'package:flutter/widgets.dart';
 
 abstract class HoverRoute extends StatelessWidget {
   String get routeName;
-  Widget regenerate(BuildContext context);
+  Widget generateRouteContent(BuildContext context);
 }