+++ lib/src/framework/framework.dart
@@ -2,3 +2,4 @@ library hover_framework;
 
 export 'hover.dart';
 export 'core/hover_page.dart';
+export 'core/components/theme/hover_theme_data.dart';